moved button examples to their own folder for cleanup purposes
