+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "radiobuttons.hpp"
+#include "buttons/radiobuttons.hpp"
 #include <gtkmm/application.h>
 
 int main( int argc, char *argv[]) {
